Remove unnecessary charset in MIME type
As referenced in the previous commit, `mime.TypeByExtension` already
includes the charset, so adding it manually is unecessary.

I also didn't add it to the "text/plain" type, since if the MIME type
can't be detected from the file extension then it's probably not safe
to assume that the file is UTF-8 encoded.